+++ src/rect.rs
@@ -2,14 +2,14 @@ use super::Point;
 
 #[derive(Clone, Copy, Debug, Default)]
 pub struct Rect {
-    pub x: isize,
-    pub y: isize,
-    pub width: usize,
-    pub height: usize,
+    pub x: i32,
+    pub y: i32,
+    pub width: u32,
+    pub height: u32,
 }
 
 impl Rect {
-    pub fn new(x: isize, y: isize, width: usize, height: usize) -> Rect {
+    pub fn new(x: i32, y: i32, width: u32, height: u32) -> Rect {
         Rect {
             x: x,
             y: y,
@@ -19,8 +19,8 @@ impl Rect {
     }
 
     pub fn contains(&self, p: Point) -> bool {
-        p.x >= self.x && p.x < self.x + self.width as isize &&
-        p.y >= self.y && p.y < self.y + self.height as isize
+        p.x >= self.x && p.x < self.x + self.width as i32 &&
+        p.y >= self.y && p.y < self.y + self.height as i32
     }
 
     pub fn point(&self) -> Point {
